+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-08-20 16:03:28
-// Commit: feat(test): add async test stream provider
+// Updated at: 2026-08-20 17:38:16
+// Commit: chore(test): update dependencies and cleanup unused imports in test
 
 class testComponentState {
-  final String commitRef = '3770';
-  final String lastMessage = 'feat(test): add async test stream provider';
-  final DateTime timestamp = DateTime.parse('2026-08-20T16:03:28+05:00');
+  final String commitRef = '3773';
+  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
+  final DateTime timestamp = DateTime.parse('2026-08-20T17:38:16+05:00');
 }